move Login page to features/auth folder
